8609-赫夫曼树：新增c.cpp修补

+++ scau-2/date-structure/8609-哈夫曼树/a.cpp
@@ -96,8 +96,8 @@ void HuffmanCoding(HuffmanTree &HT, HuffmanCode &HC, int *w, int n)
 		{
 			t = j;
 			j = HT[j].parent;
-			if(HT[j].lchild == t) p[--start] = '0';
-			if(HT[j].rchild == t) p[--start] = '1';
+			if(HT[j].lchild == (unsigned)t) p[--start] = '0';
+			if(HT[j].rchild == (unsigned)t) p[--start] = '1';
 		}
 		strcpy(HC[i], &p[start]);
 	}
